compliance_ai/requirements.txt
langgraph>=0.2.0
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-chroma>=0.4.0
anthropic>=0.40.0
chromadb>=0.5.0
streamlit>=1.40.0
pypdf>=5.0.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
tenacity>=9.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
rank-bm25>=0.2.0
numpy>=1.26.0
pandas>=2.2.0
